Renamed the BioXASCarbonFilterFarmControlView to BioXASCarbonFilterFarmView. #1915
